requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
PyJWT==2.11.0
sqlparse==0.5.5
